+++ lib/vector/diglib/portable.c
@@ -172,12 +172,12 @@ int dig__fread_port_O(off_t *buf, size_t cnt, struct gvfile *fp,
     unsigned char *c1, *c2;
 
     if (Cur_Head->off_t_quick) {
-        if (nat_off_t == port_off_t_size) {
+        if ((size_t)nat_off_t == port_off_t_size) {
             ret = dig_fread(buf, port_off_t_size, cnt, fp);
             if (ret != (int)cnt)
                 return 0;
         }
-        else if (nat_off_t > port_off_t_size) {
+        else if ((size_t)nat_off_t > port_off_t_size) {
             /* read into buffer */
             buf_alloc(cnt * port_off_t_size);
             ret = dig_fread(buffer, port_off_t_size, cnt, fp);
@@ -205,13 +205,13 @@ int dig__fread_port_O(off_t *buf, size_t cnt, struct gvfile *fp,
                 c2 += sizeof(off_t);
             }
         }
-        else if (nat_off_t < port_off_t_size) {
+        else if ((size_t)nat_off_t < port_off_t_size) {
             /* should never happen */
             G_fatal_error(_("Vector exceeds supported file size limit"));
         }
     }
     else {
-        if (nat_off_t >= port_off_t_size) {
+        if ((size_t)nat_off_t >= port_off_t_size) {
             /* read into buffer */
             buf_alloc(cnt * port_off_t_size);
             ret = dig_fread(buffer, port_off_t_size, cnt, fp);
@@ -238,7 +238,7 @@ int dig__fread_port_O(off_t *buf, size_t cnt, struct gvfile *fp,
                 c2 += sizeof(off_t);
             }
         }
-        else if (nat_off_t < port_off_t_size) {
+        else if ((size_t)nat_off_t < port_off_t_size) {
             /* should never happen */
             G_fatal_error(_("Vector exceeds supported file size limit"));
         }
@@ -640,11 +640,11 @@ int dig__fwrite_port_O(const off_t *buf, size_t cnt, struct gvfile *fp,
     unsigned char *c1, *c2;
 
     if (Cur_Head->off_t_quick) {
-        if (nat_off_t == port_off_t_size) {
+        if ((size_t)nat_off_t == port_off_t_size) {
             if (dig_fwrite(buf, port_off_t_size, cnt, fp) == cnt)
                 return 1;
         }
-        else if (nat_off_t > port_off_t_size) {
+        else if ((size_t)nat_off_t > port_off_t_size) {
             buf_alloc(cnt * port_off_t_size);
             c1 = (unsigned char *)buf;
             c2 = (unsigned char *)buffer;
@@ -660,13 +660,13 @@ int dig__fwrite_port_O(const off_t *buf, size_t cnt, struct gvfile *fp,
             if (dig_fwrite(buffer, port_off_t_size, cnt, fp) == cnt)
                 return 1;
         }
-        else if (nat_off_t < port_off_t_size) {
+        else if ((size_t)nat_off_t < port_off_t_size) {
             /* should never happen */
             G_fatal_error("Vector exceeds supported file size limit");
         }
     }
     else {
-        if (nat_off_t >= port_off_t_size) {
+        if ((size_t)nat_off_t >= port_off_t_size) {
             buf_alloc(cnt * port_off_t_size);
             c1 = (unsigned char *)buf;
             c2 = (unsigned char *)buffer;
@@ -679,7 +679,7 @@ int dig__fwrite_port_O(const off_t *buf, size_t cnt, struct gvfile *fp,
             if (dig_fwrite(buffer, port_off_t_size, cnt, fp) == cnt)
                 return 1;
         }
-        else if (nat_off_t < port_off_t_size) {
+        else if ((size_t)nat_off_t < port_off_t_size) {
             /* should never happen */
             G_fatal_error(_("Vector exceeds supported file size limit"));
         }
@@ -976,7 +976,7 @@ void dig_init_portable(struct Port_info *port, int byte_order)
     else
         port->off_t_quick = FALSE;
 
-    for (i = 0; i < nat_off_t; i++) {
+    for (i = 0; i < (size_t)nat_off_t; i++) {
         if (port->byte_order == ENDIAN_BIG)
             port->off_t_cnvrt[i] = off_t_cnvrt[i];
         else
